raspi-server/requirements.txt
blinker==1.9.0
click==8.3.0
colorzero==2.0
Flask==3.1.2
flask-cors==6.0.1
gpiozero==2.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
paho-mqtt==2.1.0
RPi.GPIO==0.7.1
Werkzeug==3.1.3
